# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "UserController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/search": {
      "get": {
        "operationId": "UserController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeSelf",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPublicDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/groups/{groupId}/ledgers": {
      "post": {
        "operationId": "LedgersController_createLedger",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLedgerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledgers"
        ]
      },
      "get": {
        "operationId": "LedgersController_listLedgers",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LedgerDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledgers"
        ]
      }
    },
    "/api/v1/ledgers/{ledgerId}": {
      "get": {
        "operationId": "LedgersController_getLedger",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledgers"
        ]
      },
      "patch": {
        "operationId": "LedgersController_updateLedger",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLedgerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledgers"
        ]
      }
    },
    "/api/v1/ledgers/{ledgerId}/participants": {
      "post": {
        "operationId": "LedgersController_addParticipant",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLedgerParticipantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledgers"
        ]
      }
    },
    "/api/v1/ledgers/{ledgerId}/participants/{userId}": {
      "delete": {
        "operationId": "LedgersController_removeParticipant",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledgers"
        ]
      }
    },
...
    "/api/v1/ledgers/{ledgerId}/settlements": {
      "post": {
        "operationId": "SettlementsController_createSettlement",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSettlementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Settlements"
        ]
      },
      "get": {
        "operationId": "SettlementsController_listSettlements",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Settlements"
        ]
      }
    },
...
  },
  "info": {
    "title": "Expensive APIs",
    "description": "The API description for the Expensive app.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "preferredCurrency": {
            "type": "string"
          },
          "preferredLocale": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "preferredCurrency",
          "preferredLocale",
          "isActive",
          "emailVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "preferredCurrency": {
            "type": "string"
          },
          "preferredLocale": {
            "type": "string"
          }
        }
      },
      "UserPublicDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "fullName"
        ]
      },
...
      "LedgerVisibility": {
        "type": "string",
        "enum": [
          "PRIVATE_TO_PARTICIPANTS",
          "VISIBLE_TO_GROUP"
        ]
      },
      "LedgerParticipantDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
      "CreateLedgerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LedgerVisibility"
              }
            ]
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerParticipantDto"
            }
          }
        },
        "required": [
          "name",
          "participants"
        ]
      },
      "LedgerParticipantPublicDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublicDTO"
          }
        },
        "required": [
          "userId",
          "joinedAt",
          "user"
        ]
      },
      "LedgerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "visibility": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LedgerVisibility"
              }
            ]
          },
          "createdById": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerParticipantPublicDto"
            }
          }
        },
        "required": [
          "id",
          "groupId",
          "name",
          "visibility",
          "createdById",
          "createdAt",
          "participants"
        ]
      },
      "UpdateLedgerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LedgerVisibility"
              }
            ]
          },
          "closedAt": {
            "type": "string"
          }
        }
      },
      "AddLedgerParticipantDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
...
      "CreateSettlementDto": {
        "type": "object",
        "properties": {
          "fromUserId": {
            "type": "string"
          },
          "toUserId": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "settlementDate": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "fromUserId",
          "toUserId",
          "amount",
          "settlementDate"
        ]
      }
    }
  }
}